# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eggpoint — B2B egg supplier directory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserSessionRole = typeof UserSessionRole[keyof typeof UserSessionRole];


export const UserSessionRole = {
  SELLER: 'SELLER',
  BUYER: 'BUYER',
  ADMIN: 'ADMIN',
} as const;

export interface UserSession {
  id: number;
  email: string;
  /** @nullable */
  phone?: string | null;
  fullName: string;
  role: UserSessionRole;
  /** @nullable */
  farmCode?: string | null;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  SELLER: 'SELLER',
  BUYER: 'BUYER',
} as const;

export interface RegisterInput {
  email: string;
  phone?: string;
  /** @minLength 2 */
  fullName: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type FarmSubscriptionTier = typeof FarmSubscriptionTier[keyof typeof FarmSubscriptionTier];


export const FarmSubscriptionTier = {
  FREE: 'FREE',
  FEATURED: 'FEATURED',
} as const;

export interface Farm {
  id: number;
  farmCode: string;
  ownerId?: number;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerPhone?: string | null;
  farmName: string;
  state: string;
  lga: string;
  /** @nullable */
  description?: string | null;
  verified: boolean;
  subscriptionTier: FarmSubscriptionTier;
  activeBatchCount?: number;
  totalCratesAvailable?: number;
  createdAt: string;
}

export interface FarmInput {
  /** @minLength 2 */
  farmName: string;
  state: string;
  lga: string;
  description?: string;
}

export interface FarmUpdate {
  /** @minLength 2 */
  farmName?: string;
  lga?: string;
  description?: string;
}

export type EggBatchEggSize = typeof EggBatchEggSize[keyof typeof EggBatchEggSize];


export const EggBatchEggSize = {
  SMALL: 'SMALL',
  MEDIUM: 'MEDIUM',
  LARGE: 'LARGE',
  JUMBO: 'JUMBO',
} as const;

export type EggBatchStatus = typeof EggBatchStatus[keyof typeof EggBatchStatus];


export const EggBatchStatus = {
  ACTIVE: 'ACTIVE',
  RESERVED: 'RESERVED',
  SOLD_OUT: 'SOLD_OUT',
  ARCHIVED: 'ARCHIVED',
} as const;

export interface EggBatch {
  id: number;
  batchCode: string;
  farmId: number;
  farmCode: string;
  farmName: string;
  farmState?: string;
  quantityCrates: number;
  eggSize: EggBatchEggSize;
  pricePerCrate: number;
  collectionDate: string;
  daysOld: number;
  status: EggBatchStatus;
  createdAt: string;
}

export type EggBatchInputEggSize = typeof EggBatchInputEggSize[keyof typeof EggBatchInputEggSize];


export const EggBatchInputEggSize = {
  SMALL: 'SMALL',
  MEDIUM: 'MEDIUM',
  LARGE: 'LARGE',
  JUMBO: 'JUMBO',
} as const;

export interface EggBatchInput {
  farmCode: string;
  /** @minimum 1 */
  quantityCrates: number;
  eggSize: EggBatchInputEggSize;
  /** @minimum 1 */
  pricePerCrate: number;
  collectionDate: string;
}

export type EggBatchUpdateStatus = typeof EggBatchUpdateStatus[keyof typeof EggBatchUpdateStatus];


export const EggBatchUpdateStatus = {
  ACTIVE: 'ACTIVE',
  RESERVED: 'RESERVED',
  SOLD_OUT: 'SOLD_OUT',
  ARCHIVED: 'ARCHIVED',
} as const;

export interface EggBatchUpdate {
  /** @minimum 1 */
  quantityCrates?: number;
  /** @minimum 1 */
  pricePerCrate?: number;
  status?: EggBatchUpdateStatus;
}

export type ListingEggSize = typeof ListingEggSize[keyof typeof ListingEggSize];


export const ListingEggSize = {
  SMALL: 'SMALL',
  MEDIUM: 'MEDIUM',
  LARGE: 'LARGE',
  JUMBO: 'JUMBO',
} as const;

export type ListingSubscriptionTier = typeof ListingSubscriptionTier[keyof typeof ListingSubscriptionTier];


export const ListingSubscriptionTier = {
  FREE: 'FREE',
  FEATURED: 'FEATURED',
} as const;

export interface Listing {
  batchCode: string;
  farmCode: string;
  farmName: string;
  state: string;
  lga: string;
  /** @nullable */
  description?: string | null;
  eggSize: ListingEggSize;
  quantityCrates: number;
  pricePerCrate: number;
  collectionDate: string;
  daysOld: number;
  verified: boolean;
  /** @nullable */
  ownerPhone: string | null;
  /** @nullable */
  ownerName?: string | null;
  subscriptionTier?: ListingSubscriptionTier;
}

export type InquiryStatus = typeof InquiryStatus[keyof typeof InquiryStatus];


export const InquiryStatus = {
  PENDING: 'PENDING',
  RESPONDED: 'RESPONDED',
  CLOSED: 'CLOSED',
} as const;

export interface Inquiry {
  id: number;
  buyerName: string;
  buyerPhone: string;
  farmCode: string;
  batchCode: string;
  quantityCrates: number;
  message: string;
  status: InquiryStatus;
  createdAt: string;
}

export interface InquiryInput {
  /** @minLength 2 */
  buyerName: string;
  /** @minLength 7 */
  buyerPhone: string;
  farmCode: string;
  batchCode: string;
  /** @minimum 1 */
  quantityCrates: number;
  /** @minLength 10 */
  message: string;
}

export type InquiryUpdateStatus = typeof InquiryUpdateStatus[keyof typeof InquiryUpdateStatus];


export const InquiryUpdateStatus = {
  PENDING: 'PENDING',
  RESPONDED: 'RESPONDED',
  CLOSED: 'CLOSED',
} as const;

export interface InquiryUpdate {
  status?: InquiryUpdateStatus;
}

export interface VerifyFarmInput {
  verified: boolean;
}

export interface PlatformStats {
  totalFarms: number;
  verifiedFarms: number;
  totalCratesAvailable: number;
  statesCovered: number;
  totalInquiries: number;
  totalBatches: number;
}

export type ListFarmsParams = {
state?: string;
search?: string;
};

export type ListBatchesParams = {
farmCode?: string;
eggSize?: string;
state?: string;
};

export type ListListingsParams = {
state?: string;
eggSize?: string;
search?: string;
minCrates?: number;
};