# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Eggpoint — B2B egg supplier directory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List egg batches
      parameters:
        - name: farmCode
          in: query
          required: false
          schema:
            type: string
        - name: eggSize
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EggBatch"
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create an egg batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EggBatchInput"
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EggBatch"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /batches/{batchCode}:
    get:
      operationId: getBatch
      tags: [batches]
      summary: Get batch detail
      parameters:
        - name: batchCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EggBatch"
        "404":
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBatch
      tags: [batches]
      summary: Update batch (status, price, quantity)
      parameters:
        - name: batchCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EggBatchUpdate"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EggBatch"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/farms:
    get:
      operationId: listPendingFarms
      tags: [admin]
      summary: List farms pending verification
      responses:
        "200":
          description: Pending farms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/farms/{farmCode}/verify:
    patch:
      operationId: verifyFarm
      tags: [admin]
      summary: Approve or reject a farm
      parameters:
        - name: farmCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyFarmInput"
      responses:
        "200":
          description: Farm verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Farm not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserSession:
      type: object
      required: [id, email, fullName, role]
      properties:
        id:
          type: integer
        email:
          type: string
        phone:
          type: ["string", "null"]
        fullName:
          type: string
        role:
          type: string
          enum: [SELLER, BUYER, ADMIN]
        farmCode:
          type: ["string", "null"]
    RegisterInput:
      type: object
      required: [email, fullName, password, role]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        fullName:
          type: string
          minLength: 2
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [SELLER, BUYER]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Farm:
      type: object
      required: [id, farmCode, farmName, state, lga, verified, subscriptionTier, createdAt]
      properties:
        id:
          type: integer
        farmCode:
          type: string
        ownerId:
          type: integer
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        farmName:
          type: string
        state:
          type: string
        lga:
          type: string
        description:
          type: ["string", "null"]
        verified:
          type: boolean
        subscriptionTier:
          type: string
          enum: [FREE, FEATURED]
        activeBatchCount:
          type: integer
        totalCratesAvailable:
          type: integer
        createdAt:
          type: string
...
    EggBatch:
      type: object
      required: [id, batchCode, farmId, farmCode, farmName, quantityCrates, eggSize, pricePerCrate, collectionDate, status, daysOld, createdAt]
      properties:
        id:
          type: integer
        batchCode:
          type: string
        farmId:
          type: integer
        farmCode:
          type: string
        farmName:
          type: string
        farmState:
          type: string
        quantityCrates:
          type: integer
        eggSize:
          type: string
          enum: [SMALL, MEDIUM, LARGE, JUMBO]
        pricePerCrate:
          type: number
        collectionDate:
          type: string
        daysOld:
          type: integer
        status:
          type: string
          enum: [ACTIVE, RESERVED, SOLD_OUT, ARCHIVED]
        createdAt:
          type: string
    EggBatchInput:
      type: object
      required: [farmCode, quantityCrates, eggSize, pricePerCrate, collectionDate]
      properties:
        farmCode:
          type: string
        quantityCrates:
          type: integer
          minimum: 1
        eggSize:
          type: string
          enum: [SMALL, MEDIUM, LARGE, JUMBO]
        pricePerCrate:
          type: number
          minimum: 1
        collectionDate:
          type: string
    EggBatchUpdate:
      type: object
      properties:
        quantityCrates:
          type: integer
          minimum: 1
        pricePerCrate:
          type: number
          minimum: 1
        status:
          type: string
          enum: [ACTIVE, RESERVED, SOLD_OUT, ARCHIVED]
...
    VerifyFarmInput:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
...