# lib/api-zod/src/generated/types/eggBatchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eggpoint — B2B egg supplier directory API
 * OpenAPI spec version: 0.1.0
 */
import type { EggBatchInputEggSize } from './eggBatchInputEggSize';

export interface EggBatchInput {
  farmCode: string;
  /** @minimum 1 */
  quantityCrates: number;
  eggSize: EggBatchInputEggSize;
  /** @minimum 1 */
  pricePerCrate: number;
  collectionDate: string;
}