# lib/api-zod/src/generated/types/userSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eggpoint — B2B egg supplier directory API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSessionRole } from './userSessionRole';

export interface UserSession {
  id: number;
  email: string;
  /** @nullable */
  phone?: string | null;
  fullName: string;
  role: UserSessionRole;
  /** @nullable */
  farmCode?: string | null;
}