# lib/api-zod/src/generated/types/verifyFarmInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eggpoint — B2B egg supplier directory API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyFarmInput {
  verified: boolean;
}